Update event tag format to match run tag.  Change format of track length seed from int to long int.

+++ Source/IO/LMCRunParameters.hh
@@ -37,7 +37,7 @@ namespace locust
             std::string fSimulationSubType;
             long int fRunID;
             long int fKassiopeiaSeed;
-            int fTrackLengthSeed;
+            long int fTrackLengthSeed;
             int fTrackDelaySeed;
 
             ClassDef(RunParameters,1)  // Root syntax.
